HO-50 Added permissions

+++ src/filesystem/src/main.c
@@ -10,6 +10,50 @@
 #include <sys/stat.h>
 #include <unistd.h>
 
+// Emscripten uses its own errno convention when compiled which
+// makes accurate error reporting convoluted - use our own.
+int translate_errors(int err) {
+  switch (err) {
+  case EEXIST:
+    return E_EXISTS;
+  case ENOENT:
+    return E_DOESNTEXIST;
+  case EACCES:
+    return E_PERM;
+  case EROFS:
+    return E_SYSFS;
+  case EISDIR:
+    return E_ISDIR;
+  case ENOTDIR:
+    return E_NOTDIR;
+  case EIO:
+    return E_IOERROR;
+  case EINVAL:
+    return E_INVALID;
+  case EAGAIN:
+    return E_AGAIN;
+  default:
+    printf("[C] Unknown error: %d\n", err);
+    return err;
+  }
+}
+
+bool can_read(struct stat *node_stat) {
+  return ((node_stat->st_mode & 0400) == 0400);
+}
+
+bool can_write(struct stat *node_stat) {
+  return ((node_stat->st_mode & 0200) == 0200);
+}
+
+bool can_exec(struct stat *node_stat) {
+  return ((node_stat->st_mode & 0100) == 0100);
+}
+
+bool is_system_file(struct stat *node_stat) {
+  return ((node_stat->st_mode & PROTECTED_BIT) == PROTECTED_BIT);
+}
+
 // Explicitly forces a filesystem synchronisation.
 // Likely not needed if the IDBFS filesystem is mounted with `autoPersist`
 // option set to TRUE
@@ -66,11 +110,52 @@ void file__initialiseFS() {
 }
 
 int file__open(const char *path, int flags, Error *err) {
-  int fd = open(path, flags);
+
+  // Permission checks
+  struct stat st;
+  bool file_exists = (stat(path, &st) == 0);
+
+  bool wants_read =
+      ((flags & O_RDONLY) == O_RDONLY) || ((flags & O_RDWR) == O_RDWR);
+  bool wants_write =
+      ((flags & O_WRONLY) == O_WRONLY) || ((flags & O_RDWR) == O_RDWR);
+
+  if (file_exists) {
+    // If the file already exists, and the user specified "create", fail
+    if ((flags & O_CREAT) == O_CREAT) {
+      *err = E_EXISTS;
+      return -1;
+    }
+    printf("We exist!");
+    // Read check
+    if (wants_read && !can_read(&st)) {
+      *err = translate_errors(EACCES);
+      return -1;
+    }
+    // Write check
+    if (wants_write && !can_write(&st)) {
+      *err = translate_errors(EACCES);
+      return -1;
+    }
+    // Trying to write to system file check
+    if (wants_write && is_system_file(&st)) {
+      *err = translate_errors(EROFS);
+      return -1;
+    }
+  } else {
+    // The file doesn't exist
+    if ((flags & O_CREAT) == 0) {
+      *err = translate_errors(ENOENT);
+      return -1;
+    }
+  }
+
+  // Finally open the file
+  int fd = open(path, flags & 0700, 0700); // Extra AND for additional defence
 
   if (fd < 0) {
     // has failed (sad) :'(
-    *err = errno;
+    *err = translate_errors(errno);
     return -1;
   }
 
@@ -79,8 +164,9 @@ int file__open(const char *path, int flags, Error *err) {
 }
 
 void file__close(int fd, Error *err) {
+  // NOTE: no perm checks, as they wouldn't make sense here
   if (close(fd) < 0) {
-    *err = errno;
+    *err = translate_errors(errno);
     return;
   }
   *err = 0;
@@ -88,10 +174,12 @@ void file__close(int fd, Error *err) {
 }
 
 void file__write(int fd, const char *content, Error *err) {
+  // NOTE: no perm checks as the user already has the file descriptor
+
   int contentLength = strlen(content);
   int written = write(fd, content, contentLength);
   if (written < 0) {
-    *err = errno;
+    *err = translate_errors(errno);
     return;
   }
   *err = 0;
@@ -100,18 +188,20 @@ void file__write(int fd, const char *content, Error *err) {
 
 // Reads up to `amt`, returning whatever it was able to read
 void file__read(int fd, int amt, ReadResult *rr, Error *err) {
+  // NOTE: No permission checks here - they already obtained fd
+
   rr->size = -1;
   rr->data = NULL;
 
   char *buf = calloc(amt, sizeof(char));
   if (!buf) {
-    *err = errno;
+    *err = translate_errors(errno);
     return;
   }
 
   int amount_read = read(fd, buf, amt);
   if (amount_read < 0) {
-    *err = errno;
+    *err = translate_errors(errno);
     return;
   }
 
@@ -125,6 +215,8 @@ void file__read(int fd, int amt, ReadResult *rr, Error *err) {
 
 // Reads the entirety of a files contents
 void file__read_all(int fd, ReadResult *rr, Error *err) {
+  // NOTE: No permission checks here - they already obtained fd
+
   rr->data = NULL;
   rr->size = -1;
 
@@ -156,7 +248,7 @@ void file__read_all(int fd, ReadResult *rr, Error *err) {
       void *blk = realloc(buf, buf_size * 2);
       if (blk == NULL) {
         // realloc failed
-        *err = errno;
+        *err = translate_errors(errno);
         free(buf);
         return;
       }
@@ -173,7 +265,7 @@ void file__read_all(int fd, ReadResult *rr, Error *err) {
   // bytes_read == 0 means EOF, if this is false, there was an error in a read
   if (bytes_read != 0) {
     // If it is not EOF
-    *err = errno;
+    *err = translate_errors(errno);
     free(buf);
     return;
   }
@@ -187,7 +279,7 @@ void file__read_all(int fd, ReadResult *rr, Error *err) {
   // Allocate enough space for the files contents into our ReadResult struct
   rr->data = malloc(rr->size * sizeof(char));
   if (!rr->data) {
-    *err = errno;
+    *err = translate_errors(errno);
     free(buf);
     return;
   }
@@ -205,9 +297,11 @@ void file__read_all(int fd, ReadResult *rr, Error *err) {
 
 // Shifts the file offset by 'amt'
 void file__shift(int fd, int amt, Error *err) {
+  // NOTE: No permission checks here - they already obtained fd
+
   int moved_bytes = lseek(fd, amt, SEEK_CUR);
   if (moved_bytes < 0) {
-    *err = errno;
+    *err = translate_errors(errno);
     return;
   }
 
@@ -217,9 +311,11 @@ void file__shift(int fd, int amt, Error *err) {
 
 // Places the file offset to `pos`
 void file__goto(int fd, int pos, Error *err) {
+  // NOTE: No permission checks here - they already obtained fd
+  
   int moved_bytes = lseek(fd, pos, SEEK_SET);
   if (moved_bytes < 0) {
-    *err = errno;
+    *err = translate_errors(errno);
     return;
   }
   *err = 0;
@@ -231,9 +327,25 @@ void file__goto(int fd, int pos, Error *err) {
 // TODO: Test on directories, if it doesnt work - likely change to
 // `file__removeFile`
 void file__remove(const char *path, Error *err) {
+  // Permission checks
+  struct stat st;
+  bool file_exists = (stat(path, &st) == 0);
+
+  if (!file_exists) {
+    *err = translate_errors(EEXIST);
+    return;
+  }
+
+  // If it's a system file, fail - user can't remove system files
+  if (is_system_file(&st)) {
+    *err = translate_errors(EROFS);
+    return;
+  }
+
+
   int error = unlink(path);
   if (error < 0) {
-    *err = errno;
+    *err = translate_errors(errno);
     return;
   }
   *err = 0;
@@ -241,9 +353,33 @@ void file__remove(const char *path, Error *err) {
 }
 
 void file__move(const char *old_path, const char *new_path, Error *err) {
+  // Permission checks [OLD FILE]
+  struct stat st;
+  bool file_exists = (stat(old_path, &st) == 0);
+
+  if (!file_exists) {
+    *err = translate_errors(EEXIST);
+    return;
+  }
+
+  // If it's a system file, fail - user can't move system files
+  if (is_system_file(&st)) {
+    *err = translate_errors(EROFS);
+    return;
+  }
+
+  // Permission checks [OLD FILE]
+  file_exists = (stat(new_path, &st) == 0);
+
+  // If it's a system file, fail - user can't overwrite system files
+  if (is_system_file(&st)) {
+    *err = translate_errors(EROFS);
+    return;
+  }
+
   int error = rename(old_path, new_path);
   if (error < 0) {
-    *err = errno;
+    *err = translate_errors(errno);
     return;
   }
   *err = 0;
@@ -251,9 +387,12 @@ void file__move(const char *old_path, const char *new_path, Error *err) {
 }
 
 void file__make_dir(const char *path, Error *err) {
+  // NOTE: No permission checks as we're not enforcing permissions
+  //       on directories.
+
   int error = mkdir(path, 0755);
   if (error < 0) {
-    *err = errno;
+    *err = translate_errors(errno);
     return;
   }
   *err = 0;
@@ -261,9 +400,12 @@ void file__make_dir(const char *path, Error *err) {
 }
 
 void file__remove_dir(const char *path, Error *err) {
+  // NOTE: No permission checks as we're not enforcing permissions
+  //       on directories.
+
   int error = rmdir(path);
   if (error < 0) {
-    *err = errno;
+    *err = translate_errors(errno);
     return;
   }
   *err = 0;
@@ -271,6 +413,9 @@ void file__remove_dir(const char *path, Error *err) {
 }
 
 void file__read_dir(const char *path, Entry *entry, int *err) {
+  // NOTE: No permission checks as we're not enforcing permissions
+  //       on directories.
+
   // Ensure previous entry name is freed before overwriting
   // (This is performed in JS API, but just to be safe!)
   if (entry->name) {
@@ -282,7 +427,7 @@ void file__read_dir(const char *path, Entry *entry, int *err) {
   if (entry->dirp == NULL) {
     entry->dirp = opendir(path);
     if (entry->dirp == NULL) {
-      *err = errno;
+      *err = translate_errors(errno);
       return;
     }
   }
@@ -301,7 +446,7 @@ void file__read_dir(const char *path, Entry *entry, int *err) {
           NULL; // Ensure `dirp` is reset to avoid reuse of closed pointer
       return;
     }
-    *err = errno;
+    *err = translate_errors(errno);
     return;
   }
 
@@ -310,7 +455,7 @@ void file__read_dir(const char *path, Entry *entry, int *err) {
   if (entry->name == NULL) {
     closedir(entry->dirp);
     entry->dirp = NULL; // Prevent reuse of closed pointer
-    *err = errno;
+    *err = translate_errors(errno);
     return;
   }
 
@@ -321,9 +466,13 @@ void file__read_dir(const char *path, Entry *entry, int *err) {
 }
 
 void file__stat(const char *name, StatResult *sr, Error *err) {
+  // NOTE: No permission checks, user can stat anything
+  // This usually relies on the `x` of the parent directory,
+  // but we're not implementing directory permissions
+
   struct stat file_stat;
   if (stat(name, &file_stat) < 0) {
-    *err = errno;
+    *err = translate_errors(errno);
     return;
   }
 
@@ -344,9 +493,13 @@ void file__stat(const char *name, StatResult *sr, Error *err) {
 }
 
 void file__fdstat(int fd, StatResult *sr, Error *err) {
+  // NOTE: No permission checks, user can stat anything
+  // This usually relies on the `x` of the parent directory,
+  // but we're not implementing directory permissions
+
   struct stat file_stat;
   if (fstat(fd, &file_stat) < 0) {
-    *err = errno;
+    *err = translate_errors(errno);
     return;
   }
 
@@ -367,8 +520,10 @@ void file__fdstat(int fd, StatResult *sr, Error *err) {
 }
 
 void file__change_dir(const char *path, Error *err) {
+  // NOTE: No permission checks here as directory permissions are ignored
+
   if (chdir(path) < 0) {
-    *err = errno;
+    *err = translate_errors(errno);
     return;
   }
 
@@ -378,28 +533,29 @@ void file__change_dir(const char *path, Error *err) {
 
 // Changes permissions (only for user - single user OS, 0[use][ignore][ignore])
 void file__permit(const char *path, int flags, Error *err) {
-  // No permissions check here - user should be allowed to modify all file permissions
-  // TODO: Figure out how to modify 
+  // permissions check
+  struct stat st;
+  bool file_exists = (stat(path, &st) == 0);
+
+  if (!file_exists) {
+    *err = translate_errors(EEXIST);
+    return;
+  }
+
+  // If it's a system file, fail
+  if (is_system_file(&st)) {
+    *err = translate_errors(EROFS);
+    return;
+  }
+
+  // No permissions check here - user should be allowed to modify all file
+  // permissions
+  // TODO: Figure out how to modify
   if (chmod(path, flags) < 0) {
-    *err = errno;
+    *err = translate_errors(errno);
     return;
   }
 
   *err = 0;
   return;
 }
-// typedef struct {
-//   int sec; // 4
-//   int nsec; // 4
-// } Time; // 8 bytes
-
-// typedef struct { // 48 bytes
-//   int size; // 4b
-//   int blocks;
-//   int blocksize;
-//   int ino;
-//   int perm; // permissions (01 Read, 010 Write) 20 bytes
-//   Time atime;
-//   Time mtime;
-//   Time ctime; // 24 bytes
-// } StatResult; // 44 bytes
